+++ backend/src/main/java/com/dh/backend/service/ProductService.java
@@ -1,55 +1,70 @@
 package com.dh.backend.service;
 
 import com.dh.backend.dto.ProductDTO;
+import com.dh.backend.exceptions.BadRequestException;
+import com.dh.backend.exceptions.ResourceNotFoundException;
 import com.dh.backend.model.Product;
 import com.dh.backend.repository.IProductRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.apache.log4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
-import java.util.HashSet;
-import java.util.List;
-import java.util.Optional;
-import java.util.Set;
+import javax.validation.constraints.NotNull;
+import java.util.*;
 
 @Service
 public class ProductService {
+
     @Autowired
-    private IProductRepository productRepository;
+    private final IProductRepository productRepository;
+
+    public ProductService(IProductRepository productRepository) {
+        this.productRepository = productRepository;
+    }
+
+    @Autowired
+    private CityService cityService;
+
+    private static final Logger logger = Logger.getLogger(ProductService.class);
 
     @Autowired
     ObjectMapper mapper;
-
-
-    /**
-     * Métodos CRUD completos + Método listar
-     */
 
     /**
      * Crear
      * @param productDTO
      * @return Graba en BBDD y retorna un DTO
      */
-    public ProductDTO createProduct(ProductDTO productDTO) {
+    public ProductDTO createProduct(ProductDTO productDTO) throws BadRequestException {
+        if(productDTO.getName().isEmpty() || productDTO == null)
+            throw new BadRequestException("El producto no puede ser null");
+        Long cityId = productDTO.getCity_id().getId();
         Product product = mapper.convertValue(productDTO, Product.class);
         if (this.findProductByName(product.getName()) != null)
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Este producto ya existe");
+            throw new BadRequestException("Este producto ya existe");
 
-        return mapper.convertValue(productRepository.save(product), ProductDTO.class);
+        productDTO = mapper.convertValue(productRepository.save(product), ProductDTO.class);
+        productDTO.setCity_id(cityService.findCityById(cityId));
+
+        logger.info("Categoria creada exitosamente");
+
+        return productDTO;
     }
+
 
     /**
      * Buscar por id
      * @param id
      * @return Retorna el DTO que corresponde a ese ID
      */
-    public ProductDTO readProduct(Long id) {
+    public ProductDTO readProduct(Long id) throws BadRequestException, ResourceNotFoundException {
+        if( id == null || id < 1 )
+            throw new BadRequestException("El id del producto no puede ser null ni negativo");
         if (productRepository.findById(id).isEmpty())
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No existe el producto con el id: " + id);
-
+            throw new ResourceNotFoundException("No existe existe  el producto con id: " + id);
         Optional<Product> product = productRepository.findById(id);
+        logger.info("Se encontró la categoría " + product.get().getName());
         return mapper.convertValue(product, ProductDTO.class);
     }
 
@@ -58,22 +73,42 @@ public class ProductService {
      * @param productDTO
      * @return Graba cambios en BBDD y retorna el DTO
      */
-    public ProductDTO updateProduct(ProductDTO productDTO) {
-        Product product = mapper.convertValue(productDTO, Product.class);
-        if (productRepository.findById(product.getId()).isEmpty())
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No existe el producto que quieres modificar");
+    public ProductDTO updateProduct(ProductDTO productDTO) throws BadRequestException, ResourceNotFoundException {
 
-        return mapper.convertValue(productRepository.save(product), ProductDTO.class);
+
+        if (productDTO == null )
+            throw new BadRequestException("El producto no puede ser null");
+        if (productDTO.getId() == null)
+            throw new BadRequestException("El id del producto no puede ser null");
+        Product product = mapper.convertValue(productDTO, Product.class);
+        Product productUpdate = productRepository.findById(product.getId()).orElse(null);
+        if (productUpdate == null)
+            throw new ResourceNotFoundException("No existe el producto que quieres modificar");
+        productUpdate.setName(product.getName());
+        productUpdate.setDescription(product.getDescription());
+        productUpdate.setAvailability(product.getAvailability());
+        productUpdate.setPolicies(product.getPolicies());
+        productUpdate.setLongitude(product.getLongitude());
+        productUpdate.setLatitude(product.getLatitude());
+        productUpdate.setImages(product.getImages());
+        productUpdate.setCity_id(product.getCity_id());
+        productUpdate.setCategory_id(product.getCategory_id());
+        productUpdate.setCharacteristicsInProducts_id(product.getCharacteristicsInProducts_id());
+
+        logger.info("Producto modificado exitosamente");
+        return mapper.convertValue(productRepository.save(productUpdate), ProductDTO.class);
     }
 
     /**
      * Eliminar
      * @param id Elimina según id
      */
-    public void deleteProduct(Long id) {
+    public void deleteProduct(Long id) throws BadRequestException, ResourceNotFoundException {
+        if( id == null || id < 1 )
+            throw new BadRequestException("El id del producto no puede ser null ni negativo");
         if (productRepository.findById(id).isEmpty())
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "No existe el producto con el id: " + id);
-
+            throw new ResourceNotFoundException("No existe el producto con id: " + id);
+        logger.info("Producto eliminado exitosamente");
         productRepository.deleteById(id);
     }
 
@@ -82,13 +117,66 @@ public class ProductService {
      * @return Retorna listado completo de entidades
      */
     public Set<ProductDTO> getListProduct() {
-        List<Product> categories = productRepository.findAll();
-        Set<ProductDTO> categoriesDTO = new HashSet<>();
-        for (Product product: categories) {
-            categoriesDTO.add(mapper.convertValue(product, ProductDTO.class));
+        List<Product> products = productRepository.findAll();
+        Set<ProductDTO> productsDTO = new HashSet<>();
+        for (Product product: products) {
+            productsDTO.add(mapper.convertValue(product, ProductDTO.class));
         }
 
-        return categoriesDTO;
+        return productsDTO;
+    }
+
+    // Lista de Productos según ciudad
+    public List<ProductDTO> findProductsByCity(Long id)  throws BadRequestException {
+        if (id == null || id < 1)
+            throw new BadRequestException("El id de la ciudad no puede ser null, ni negativo");
+
+        List<Product> products = productRepository.findProductsByCity(id);
+        List<ProductDTO> productsDTO = new ArrayList<>();
+        for (Product product: products) {
+            productsDTO.add(mapper.convertValue(product, ProductDTO.class));
+        }
+
+        return productsDTO;
+    }
+
+
+    // Lista de Productos según categoría
+    public List<ProductDTO> findProductsByCategory(Long id) throws BadRequestException {
+        if (id == null || id < 1)
+            throw new BadRequestException("El id de la categoría no puede ser null, ni negativo");
+
+        List<Product> products = productRepository.findProductsByCategory(id);
+        List<ProductDTO> productsDTO = new ArrayList<>();
+        for (Product product: products) {
+            productsDTO.add(mapper.convertValue(product, ProductDTO.class));
+        }
+        return productsDTO;
+    }
+
+
+
+    // Lista de Productos Random
+    public List<ProductDTO> getProductsRandom() {
+        int NUMBER_OF_ELEMENTS = 3;
+        Random rand = new Random();
+
+        List<Product> products = productRepository.findAll();
+        List<ProductDTO> productsDTO = new ArrayList<>();
+        List<Product> productsRandom = new ArrayList<>();
+
+        for (int i = 0; i < NUMBER_OF_ELEMENTS; i++) {
+            int randomIndex = rand.nextInt(products.size());
+            Product randomElement = products.get(randomIndex);
+            productsRandom.add(randomElement);
+            products.remove(randomElement);
+        }
+
+        for (Product product: productsRandom) {
+            productsDTO.add(mapper.convertValue(product, ProductDTO.class));
+        }
+
+        return productsDTO;
     }
 
     /**.
@@ -99,5 +187,4 @@ public class ProductService {
     public Product findProductByName(String name) {
         return productRepository.findProductByName(name);
     }
-
 }
